add menu wrapper
